// superlista-front/src/openapi/model/categoriaIngredienteDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoriaIngredienteDto { 
    id?: number;
    nombre?: string;
    descripcion?: string;
}